Reapply "add maven"
This reverts commit 13d7aa30f6982178bdc3082310e73f75602cc452.
