--- src/Main.java
@@ -1,14 +1,0 @@
-import java.io.IOException;
-
-public class Main {
-    public static void main(String[] args) throws IOException {
-
-
-
-        Menu.menu();
-
-
-
-    }
-
-}
