+++ src/main/java/org/vederko/Main.java
@@ -1,0 +1,10 @@
+package main.java.org.vederko;
+
+import java.io.IOException;
+
+public class Main {
+
+    public static void main(String[] args) throws IOException {
+        Menu.menu();
+    }
+}
