+++ src/main/java/org/vederko/model/Man.java
@@ -1,0 +1,16 @@
+package main.java.org.vederko.model;
+
+public class Man extends Driver {
+    private int footSize;
+
+    public void setFootSize(int footSize) {
+        this.footSize = footSize;
+    }
+
+    public int getFootSize() {
+
+        return footSize;
+    }
+
+
+}
