+++ src/main/java/org/vederko/model/Passenger.java
@@ -1,0 +1,16 @@
+package main.java.org.vederko.model;
+
+public class Passenger extends Car {
+
+    private int numberOfPassengerSeats;
+
+    public void setNumberOfPassengerSeats(int numberOfPassengerSeats) {
+        this.numberOfPassengerSeats = numberOfPassengerSeats;
+    }
+
+    public int getNumberOfPassengerSeats() {
+
+        return numberOfPassengerSeats;
+    }
+
+}
