+++ src/main/java/org/vederko/model/People.java
@@ -1,0 +1,15 @@
+package main.java.org.vederko.model;
+
+public class People extends Driver {
+    private int height;
+
+    public void setHeight(int height) {
+        this.height = height;
+    }
+
+    public int getHeight() {
+
+        return height;
+    }
+
+}
