+++ src/main/java/org/vederko/model/Truck.java
@@ -1,0 +1,16 @@
+package main.java.org.vederko.model;
+
+public class Truck extends Car {
+
+    private int loadCapacity;
+
+    public void setLoadCapacity(int loadCapacity) {
+        this.loadCapacity = loadCapacity;
+    }
+
+    public int getLoadCapacity() {
+
+        return loadCapacity;
+    }
+
+}
