+++ src/main/java/org/vederko/model/Woman.java
@@ -1,0 +1,13 @@
+package main.java.org.vederko.model;
+
+public class Woman extends Driver {
+    private int bustSize;
+
+    public void setBustSize(int bustSize) {
+        this.bustSize = bustSize;
+    }
+
+    public int getBustSize() {
+        return bustSize;
+    }
+}
