+++ src/main/java/org/vederko/model/enums/Brand.java
@@ -1,0 +1,23 @@
+package main.java.org.vederko.model.enums;
+
+public enum Brand {
+
+    AUDI("AUDI"),
+
+    MERCEDES("MERCEDES"),
+
+    BMW("BMW");
+
+    private String str;
+
+    Brand(String str) {
+
+        this.str = str;
+    }
+
+    public String getStr(){
+
+        return str;
+    }
+
+}
