+++ src/main/java/org/vederko/model/enums/CategoryOfRights.java
@@ -1,0 +1,23 @@
+package main.java.org.vederko.model.enums;
+
+public enum CategoryOfRights {
+    A("A"),
+
+    B("B"),
+
+    C("C");
+
+    private String str;
+
+    CategoryOfRights(String str) {
+
+        this.str = str;
+    }
+
+    public String getStr(){
+
+        return str;
+    }
+
+
+}
