+++ src/main/java/org/vederko/model/enums/Color.java
@@ -1,0 +1,22 @@
+package main.java.org.vederko.model.enums;
+
+public enum Color {
+    YELLOW("yellow"),
+
+    BLUE("blue"),
+
+    BLACK("black");
+
+    private String str;
+
+    Color(String str) {
+        this.str = str;
+    }
+
+    public String getStr(){
+        return str;
+    }
+
+
+
+}
