--- src/model/Man.java
@@ -1,16 +1,0 @@
-package model;
-
-public class Man extends Driver {
-    private int footSize;
-
-    public void setFootSize(int footSize) {
-        this.footSize = footSize;
-    }
-
-    public int getFootSize() {
-
-        return footSize;
-    }
-
-
-}
