--- src/model/Passenger.java
@@ -1,16 +1,0 @@
-package model;
-
-public class Passenger extends Car {
-
-    private int numberOfPassengerSeats;
-
-    public void setNumberOfPassengerSeats(int numberOfPassengerSeats) {
-        this.numberOfPassengerSeats = numberOfPassengerSeats;
-    }
-
-    public int getNumberOfPassengerSeats() {
-
-        return numberOfPassengerSeats;
-    }
-
-}
