--- src/model/People.java
@@ -1,15 +1,0 @@
-package model;
-
-public class People extends Driver {
-    private int height;
-
-    public void setHeight(int height) {
-        this.height = height;
-    }
-
-    public int getHeight() {
-
-        return height;
-    }
-
-}
