--- src/model/Truck.java
@@ -1,16 +1,0 @@
-package model;
-
-public class Truck extends Car {
-
-    private int loadCapacity;
-
-    public void setLoadCapacity(int loadCapacity) {
-        this.loadCapacity = loadCapacity;
-    }
-
-    public int getLoadCapacity() {
-
-        return loadCapacity;
-    }
-
-}
