--- src/model/Woman.java
@@ -1,13 +1,0 @@
-package model;
-
-public class Woman extends Driver {
-    private int bustSize;
-
-    public void setBustSize(int bustSize) {
-        this.bustSize = bustSize;
-    }
-
-    public int getBustSize() {
-        return bustSize;
-    }
-}
